Preprocessing - all maturities cleaned
Had to rewrite cleaning into the loading up process, so it's smaller straight away. With the unclean version I couldn't clean it due to it being already 25GB and any operation on it was unsuccessful (datetime edits etc). Now we have all maturities in one object and they're already cleaned up.