Models - factor custom Nelson.Siegel function
Optimised the function Nelson.Siegel_custom_lambda_parallel() so it is parallelised code, and used the package "parallel" to calculate on all CPU cores. GPU packages for the lm(), like gpuLm(), are not working now, deprecated this one. It runs fast enough on all CPU cores, from 1h to 13min.